feature : 변수 리네임 및 로직 AWS 서비스로 이동
